+++ src/utils/database-common/src/db_credentials.rs
@@ -7,101 +7,14 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-use std::path::Path;
-
-use secrecy::{ExposeSecret, Secret};
-
-use crate::DatabaseProvider;
+use secrecy::Secret;
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 #[derive(Debug, Clone)]
 pub struct DatabaseCredentials {
-    pub provider: DatabaseProvider,
-    pub user: String,
-    pub database_name: String,
-    pub host: String,
-    pub port: Option<u32>,
-}
-
-impl DatabaseCredentials {
-    pub fn new(
-        provider: DatabaseProvider,
-        user: String,
-        database_name: String,
-        host: String,
-        port: Option<u32>,
-    ) -> Self {
-        Self {
-            provider,
-            user,
-            database_name,
-            host,
-            port,
-        }
-    }
-
-    pub fn port(&self) -> u32 {
-        self.port.unwrap_or_else(|| self.provider.default_port())
-    }
-
-    pub fn connection_string(&self, password: Option<Secret<String>>) -> Secret<String> {
-        if let DatabaseProvider::Sqlite = self.provider {
-            Secret::new(format!("{}://{}", self.provider, self.database_name))
-        } else if let Some(password) = password {
-            Secret::new(format!(
-                "{}://{}:{}@{}:{}/{}",
-                self.provider,
-                self.user,
-                password.expose_secret(),
-                self.host,
-                self.port(),
-                self.database_name
-            ))
-        } else {
-            Secret::new(format!(
-                "{}://{}@{}:{}/{}",
-                self.provider,
-                self.user,
-                self.host,
-                self.port(),
-                self.database_name
-            ))
-        }
-    }
-
-    pub fn connection_string_no_db(&self, password: Option<Secret<String>>) -> Secret<String> {
-        if let DatabaseProvider::Sqlite = self.provider {
-            panic!("Sqlite does not support connection strings without DB")
-        } else if let Some(password) = password {
-            Secret::new(format!(
-                "{}://{}:{}@{}:{}",
-                self.provider,
-                self.user,
-                password.expose_secret(),
-                self.host,
-                self.port(),
-            ))
-        } else {
-            Secret::new(format!(
-                "{}://{}@{}:{}",
-                self.provider,
-                self.user,
-                self.host,
-                self.port(),
-            ))
-        }
-    }
-
-    pub fn sqlite_from(path: &Path) -> Self {
-        Self {
-            provider: DatabaseProvider::Sqlite,
-            user: String::new(),
-            database_name: String::from(path.to_str().unwrap()),
-            host: String::new(),
-            port: None,
-        }
-    }
+    pub user_name: Secret<String>,
+    pub password: Secret<String>,
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////